[RFC] shell: allow overriding built-in commands
This patch allows overriding built-in commands by placing a script
with the same name under git-shell-commands directory.

This is useful for users who want to extend the built-in commands
without replacing the original command binary. For instance, a user
wanting to allow only a subset of users to run the git-receive-pack
can override the command with a script that checks the user and
calls the original command if the user is allowed.

CC: Junio C Hamano <gitster@pobox.com>, Jeff King <peff@peff.net>
CC: Taylor Blau <me@ttaylorr.com>
Signed-off-by: Ayman Bagabas <ayman.bagabas@gmail.com>