git: add --no-hooks global option
Git has several hooks which are executed during certain events as long
as those hooks are enabled in the hooks directory (possibly moved from
.git/hooks via the core.hooksPath config option). These are configured
by the user, or perhaps by tooling the user has agreed to, and are not
required to operate a Git repository.

In some situations, these hooks have poor performance and expert users
may want to skip the hooks as they don't seem to affect the current
situation. One example is a pre-commit hook that checks for certain
structures in the local changes, but expert users are likely to have
done the right thing in advance.

I have come across users who have disabled hooks themselves either by
deleting hooks (supported, safe) or setting 'core.hooksPath' to some
bogus path (seems unsafe). The supported process is painful to swap
between the hook-enabled scenario and the hook-disabled scenario.

To that end, add a new --no-hooks global option to allow users to
disable hooks quickly. This option is modeled similarly to the
--no-advice option in b79deeb554 (advice: add --no-advice global option,
2024-05-03). This uses a GIT_HOOKS environment variable to communicate
to subprocesses as well as making this a backwards-compatible way for
tools to signal that they want to disable hooks.

The critical piece is that all hooks pass through run_hooks_opt() where
a static int will evaluate the environment variable and store that the
variable is initialized for faster repeated runs.

Signed-off-by: Derrick Stolee <stolee@gmail.com>