ci(win+Meson): build in Release mode, avoiding t7001-mv hangs
Since switching to `--vsenv`, the t7001-mv test consistently times out
after six hours in the CI builds on GitHub. This kind of waste is
inconsistent with my values.

The reason for this timeout is the test case 'nonsense mv triggers
assertion failure and partially updated index' in t7001-mv (which is
not even a regression test, but instead merely demonstrates a bug that
someone thought someone else should fix at some time). As the name
suggests, it triggers an assertion. The problem with this is that an
assertion on Windows, at least when run in Debug mode, will open a modal
dialog that patiently awaits some buttons to be clicked. Which never
happens in automated builds.

The solution is straight-forward: Just like the `win+VS` job already did
in forever, build in Release mode (where that modal assertion dialog is
never shown).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>