config: read from config sequence for global scope
Based on
https://lore.kernel.org/git/kl6ly1oze7wb.fsf@chooglen-macbookpro.roam.corp.google.com.

We assumed that git config is always read from a single file, but that
is not the case for the global scope, which takes inputs from both
`$HOME/.gitconfig` and `$XDG_CONFIG_HOME/git/config`. The existing
`do_git_config_sequence()` function respects both locations, in addition
to other scopes. Hence, modify the function to support reading only the
global scope, by introducing new flags that exclude everything else
(i.e. `ignore_system` and `ignore_global`).

I've assumed that `config_source->scope == CONFIG_SCOPE_GLOBAL` iff
`--global` is specified. This allows us to call the modified
`do_git_config_sequence()` function when reading only the the global
scope, rather than calling
`git_config_from_file_with_options(..., "~/.gitconfig", ...)`. This
fixes the issue where `XDG_CONFIG_HOME` location is ignored in
`git config list --global`.

NB: The `ignore_global` flag is not set to truthy anywhere, so the
`if (!opts->ignore_global)` condition is always met. We can actually
remove this unused flag if we want.

The `opts->source.file` is still set in `builtin/config.c` because it is
required for write operations. I thought that this would be misleading
because there is no single source of truth for the source files in the
global scope. So I added comments that will (hopefully) clarify this.

Reported-by: Jade Lovelace <lists@jade.fyi>
Suggested-by: Glen Choo <glencbz@gmail.com>
Signed-off-by: Delilah Ashley Wu <delilahwu@microsoft.com>