pack-bitmap: reword comments in test_bitmap_commits()
In pack-bitmap.c:test_bitmap_commits(), it comments

    /*
     * As this function is only used to print bitmap selected
     * commits, we don't have to read the commit table.
     */

This suggests that we can avoid reading the commit table altogether.
However, this comment is misleading. The reason we load bitmap entries here
is because test_bitmap_commits() needs to print the commit IDs from the
bitmap, and we must read the bitmap entries to obtain those commit IDs.
So reword this comment.

Signed-off-by: Lidong Yan <502024330056@smail.nju.edu.cn>