doc: git-log: convert pretty formats to new doc format
- Use _<placeholder>_ instead of <placeholder> in the description
- Use `backticks` for keywords and more complex option
descriptions. The new rendering engine will apply synopsis rules to
these spans.

For all the formats in the form of %(foo), the formatting needs to be heavier
because we not want the parentheses to be rendered as syntax elements,but as
keywords, i.e. we need to circumvent the syntax highlighting of synopsis.
In this particular case, this requires the heavy escaping of the parts that
contain parentheses with ++.

Signed-off-by: Jean-Noël Avila <jn.avila@free.fr>