daemon: use sigaction() to install child_handler()
In a future change, the flags used for processing SIGCHLD will need to
be updated, which is only possible by using sigaction().

Factor out the call to set the signal handler and use sigaction instead
of signal for the systems that allow that, which has the added benefit
of using BSD semantics reliably and therefore not needing the rearming
call.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>