daemon: explicitly allow EINTR during poll()
If the setup for the SIGCHLD signal handler sets SA_RESTART, poll()
might not return with -1 and set errno to EINTR when a signal is
received.

Since the logic to reap zombie childs relies om those interruptions
make sure to explicitly disable SA_RESTART around this function.

Add a Makefile flag for portability to systems that don't have the
functionality to change those flags or where it is not needed.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>