docs: correct mistakes in git-merge
Documentation for git-merge incorrectly notes that
tip of the current branch on ascii diagram is C
, while it is actually G (current branch is
master, HEAD on diagram is G).

Additionally diagrams on the page are adjusted
to use spaces instead of tabs, so that they align
regardless of tab size. This is in line with
diagrams on other git documentation pages.

Signed-off-by: Timur Sultanaev <str.write@gmail.com>