sparse-checkout: add 'clean' command
When users change their sparse-checkout definitions to add new
directories and remove old ones, there may be a few reasons why
directories no longer in scope remain (ignored or excluded files still
exist, Windows handles are still open, etc.). When these files still
exist, the sparse index feature notices that a tracked, but sparse,
directory still exists on disk and thus the index expands. This causes a
performance hit _and_ the advice printed isn't very helpful. Using 'git
clean' isn't enough (generally '-dfx' may be needed) but also this may
not be sufficient.

Add a new subcommand to 'git sparse-checkout' that removes these
tracked-but-sparse directories, including any excluded or ignored files
underneath. This is the most extreme method for doing this, but it works
when the sparse-checkout is in cone mode and is expected to rescope
based on directories, not files.

Be sure to add a --dry-run option so users can predict what will be
deleted. In general, output the directories that are being removed so
users can know what was removed.

Note that untracked directories remain. Further, directories that
contain staged changes are not deleted. This is a detail that is partly
hidden by the implementation which relies on collapsing the index to a
sparse index in-memory and only deleting directories that are listed as
sparse in the index. If a staged change exists, then that entry is not
stored as a sparse tree entry and thus remains on-disk until committed
or reset.

Signed-off-by: Derrick Stolee <stolee@gmail.com>